Update references 2.0

+++ ms-ticket/src/main/java/com/proyectointegrador/msticket/repository/ITicketRepository.java
@@ -3,5 +3,5 @@ package com.proyectointegrador.msticket.repository;
 import com.proyectointegrador.msticket.domain.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ITicketRepository extends JpaRepository<Ticket, Integer> {
+public interface ITicketRepository extends JpaRepository<Ticket, Long> {
 }
